backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
PyMuPDF==1.27.1
httpx==0.28.1
Pillow==11.1.0
google-genai==1.5.0
sse-starlette==2.2.1
pydantic==2.10.5
pydantic-settings==2.7.1
aiofiles==24.1.0
python-docx==1.2.0
beautifulsoup4==4.14.3
playwright==1.58.0
pdf2docx==0.5.10
doclayout-yolo==0.0.4

